Fix some errors when selecting empty result
Selectors with a select callback which didn't check if the result was
empty could fail with errors like "index out of range", "key not in
dictionary", because Fuzzbox invoked the callback with an empty result.

Fix by checking result is not empty before invoking select callback.
This is a special case for invoking a callback or other action, we know
there is no point in invoking the select action for an empty result. The
same does not apply to other actions, so this only applies to select.