Restore support for selecting empty result
I was too eager when removing this as a fix all solution to errors in
selectors that didn't allow for an empty result. Leave in place until
absolutely sure that support is not necessary, and removing it is not
confusing (could be, as select callbacks would be invoked differently
to other callbacks and actions, conditionally, only with a selection).