Fix issue clicking on scrollbar if enabled
There is no general option to enable a scrollbar, but it can be enabled
for a selector by setting `scrollbar: true` in opts to selector.Start()

While the scrollbar would appear, and move as expected with mouse wheel
actions, clicking on it didn't behave as documented in the Vim popup
help as left mouse click was already being captured in the menu filter.

This just tweaks left mouse handling to allow for clicking on scrollbar.