Tidy code to apply devicons offset to highlight
Apply this once, in selector.UpdateMenu(), to avoid confusing duplicated
code and to remove need for selectors to know anything about devicons.

At the moment this probably slightly degrades performance as it now
processes the async results twice, once to convert to separate lists of
matched strings and highlight positions, and again to apply the offset.
I'm not worrying about that too much for now, preferring to tidy first
and find performance optimisations later (easier with cleaner code).

This also updates the function signature for the FuzzySearchAsync
callback to be consistent with args the FuzzySearch function returns.