Jump to existing terminal buffer win when opening
When opening a "file" that is a special buffer, i.e. not a file that can
be edited, most likely a terminal buffer found from listed buffers, jump
to the buffer in the current tab if it is visible in a widow.

This is not consistent with opening readable files, which are always
opened in the current window or the next usable window, but it avoids
accidentally loading the same terminal buffer into multiple windows.