Add g:fuzzbox_scrollbar, enable scrollbar in menu
Makes most sense combined with disabling preview. There is no option to
enable scrollbars in both the menu and the preview, mostly because it
requires a lot of hacking to ensure the window is correctly sized for
the scrollbar and I can't see much value in a scrollbar in the preview.