Tweak triggering menu updates from FuzzyPrevious
Rather than calling feedkeys() to trigger the menu update, just load the
full previous search into the prompt on launch and again after 100ms -
SetPrompt() function triggers the input callback, and updates the menu.

This is still just a hack, completely reliant on knowledge of how both
the files and grep selectors work, initially launching with no items in
the selection list and then updating the menu with the current result of
a background job after 100ms. Much tidying up remains to be done here!