Make FuzzyFiles play nicer with FuzzyPrevious
Update files selector to reset the prompt and trigger the input callback
after the background job to get the list of files for selection exits,
and remove hack from launcher that did the same after a 100ms delay.

The hack in the launcher mostly worked, but it was a horrible hack, it's
really the responsibility of the selector to handle the input, wherever
it comes from or when. The hack also didn't work when the file command
didn't respond with output within 100ms, e.g. using standard Unix find
on a codebase with hundreds of thousands of files (sure, why would you,
it's a bad idea, but Fuzzbox should still work, even very slowly).