Add jumps selector and FuzzyJumps command
See :h jumplist and :h :jumps

Like Telescope and FZF.vim this shows the list of jumps as a list of
files, with text snippet if available. This is not how Vim itself shows
the jumplist in the output from the :jumps command, so it's not as clear
how many jumps are required to move to a particular position in the
list, or quite so clear which jumps are in the current buffer, but it
works well enough, and might even be more appropriate for fuzzy finding.