Add jumps selector and FuzzyJumps command
This doesn't quite work how I'd like, but it does work and has been
sitting in a branch for about 6 weeks, so it seems worth merging.

Shows the jump list as a list of files, which is similar to FZF.vim and
Telescope.nvim. Allows those files to be opened at the specified line
and column, but doesn't currently work well with [No Name] buffers
(neither do FZF.vim or Telescope.nvim). I'd prefer to improve this by
using the output from the :jumps command, but previewing is tricky.