Fix FuzzyFiles glitch when clearing prompt
Make sure to stop any running async timer when the input in the prompt
is deleted. Without this is was possible for a running timer to complete
after initially updating the menu, so you could see no input in the
prompt, but a pattern matched in the menu.

This was only really noticeable when searching hundreds of thousands of
files, and deleting individual characters with backspace, otherwise the
async worker would complete fast enough not to override the menu.

Note: I use a custom command to search all files using `rg -uu --files`