Rename utils directory to internal
These are the scripts that make up the core of Fuzzbox, but they are not
to be considered public, do not use them as an api to build extensions
(hence the new name, internal = here be dragons, proceed with caution!).

The files directly within autoload/fuzzbox/ make up the public api for
building extensions. At the moment quite limited, but sufficient for
simple extensions. More files and functions will be added over time.