Simplify code to update results in files selector
The files selector uses a background job to compile a list of files, and
asynchronously searches that list, sometimes while the list is being
compiled. This means there are quite a few scenarios for updating the
menu and counter, and the code had gotten very hacky dealing with them.

This commit tidies things up, without changing behaviour, by modifying
the UpdateMenu() function to allow it be called in response to input and
when the background job finishes, thereby removing the need for custom
code within input and job exit handlers to update the menu and counter.