Fix menu highlight when using Unix find
Broken by 81d7000e8d05e53497fddcdabf1527cf3371b14e, which tidied up
devicons handling, but clearly was not tested sufficiently. The root
cause of this bug was the hack to fix Windows file separators on Unix
emulation environments when processing results from the background job
to build the list of files (it removed the leading ./ from relative
paths). The fix for the bug is a better solution to that problem (don't
fix the paths when rendering the results in the menu, only when trying
to read the files - this should also improve performance a little).