Code tidy, files and grep command builders
Move files cmdbuilder script to files subdirectory of builtins, because
it is a dependency of the files selector, not a general command builder.

Extract grep cmdbuilder following similar pattern. This is a little odd
as the builder function doesn't return a command to run but a template
string, format pattern to match results, and ignore case option to be
used when mimicking smart case for grep programs that don't support it.
Will probably revisit and refactor, but it'll do for now.