Detect preview filetype without triggering autocmds
This avoids triggering FileType autocmds when detecting the filetype for
the preview window, which is done to apply the same syntax highlighting
to the preview that will be applied when the file is opened for editing.

Notably this addresses compatibility issues with LSP plugins as they
tend to rely on the FileType autocmd event to automatically start
language servers (though yegappan/lsp now skips this for popup wins).

Mostly this works the exact same as using Vim's filetypedetect autocmds
as it copies many commands from those autocmds into a new autocmd group,
but not everything can be copied, so there will be some discrepancies.