Improve filetype detection, also match setfiletype
Copy setfiletype commands in addition to setf abbreviation (though setf
is always use in Vim's built in filetype detection, plugins or custom
filetype detection in a users' vimrc might use setfiletype).

This still only handles simple `setf|setfiletype <filetype>` commands,
but that should catch most well-behaved plugins and vimrc customisation.
It is of course possible to `set filetype=foo`, but not recommended as
doing so does not check if the filetype has already been detected.